requirements.txt
streamlit
torch
torchvision
segmentation-models-pytorch
opencv-python-headless
numpy
pandas
Pillow
joblib
scikit-learn
gdown
